requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
asgiref==3.7.2
babel==2.16.0
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==5.0.6
django-admin-interface==0.28.6
django-bootstrap5==24.2
django-colorfield==0.11.0
django-extra-settings==0.12.0
django-import-export==4.1.0
django-phonenumber-field==7.3.0
django-pwa==2.0.0
django-qr-code==4.1.0
django-widget-tweaks==1.4.12
docutils==0.20.1
et-xmlfile==1.1.0
idna==3.7
imagesize==1.4.1
Jinja2==3.1.4
jsonfield==3.1.0
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.2.1
phonenumbers==8.13.30
pillow==10.3.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.32.3
segno==1.6.1
setuptools==73.0.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-markdown-builder==0.6.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.4.4
tablib==3.5.0
tabulate==0.9.0
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.2
xlrd==2.0.1
